container/requirements.txt
aioflask
gunicorn
pandas
numpy
requests
nest-asyncio
pyppeteer
pyppdf
absl-py
google-cloud-storage
google-cloud-logging
cloud-sql-python-connector
pg8000
sqlalchemy
bs4